+++ quant_strategy_tokenizer/hash_v2/__init__.py
@@ -7,7 +7,10 @@ from quant_strategy_tokenizer.hash_v2.graph_hash import graph_hash_v2
 from quant_strategy_tokenizer.hash_v2.implementation_ref_hash import implementation_ref_hash_v2
 from quant_strategy_tokenizer.hash_v2.instance_hash import compute_hashes_v2, instance_hash_v2
 from quant_strategy_tokenizer.hash_v2.param_hash import param_hash_v2
-from quant_strategy_tokenizer.hash_v2.signature_hash import signature_hash_v2
+from quant_strategy_tokenizer.hash_v2.signature_hash import (
+    signature_hash_for_ports_v2,
+    signature_hash_v2,
+)
 from quant_strategy_tokenizer.hash_v2.token_pack_hash import token_pack_hash_v2
 from quant_strategy_tokenizer.hash_v2.token_spec_hash import token_spec_hash_v2
 
@@ -22,6 +25,7 @@ __all__ = [
     "implementation_ref_hash_v2",
     "instance_hash_v2",
     "param_hash_v2",
+    "signature_hash_for_ports_v2",
     "signature_hash_v2",
     "token_pack_hash_v2",
     "token_spec_hash_v2",